requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
deezer-oauth-cli==0.2.5
idna==3.3
Pygments==2.13.0
requests==2.28.1
rich==12.5.1
shellingham==1.5.0
typer==0.6.1
urllib3==1.26.11
